requirements.txt
streamlit==0.87.0
mlflow ==  1.17.0
pandas == 1.2.4
pillow == 8.0.1
sqlalchemy == 1.4.7
altair == 4.1.0
pycaret == 2.3.3
